+++ src/helper/png-write.h
@@ -16,9 +16,9 @@
 struct SPDocument;
 
 bool sp_export_png_file (SPDocument *doc, gchar const *filename,
-			 double x0, double y0, double x1, double y1,
-			 unsigned int width, unsigned int height, double xdpi, double ydpi,
-			 unsigned long bgcolor,
-			 unsigned int (*status) (float, void *), void *data, bool force_overwrite = false, GSList *items_only = NULL);
+            double x0, double y0, double x1, double y1,
+            unsigned long int width, unsigned long int height, double xdpi, double ydpi,
+            unsigned long bgcolor,
+            unsigned int (*status) (float, void *), void *data, bool force_overwrite = false, GSList *items_only = NULL);
 
 #endif
